Created Simplified Class for `System.Security.Claims.Claim`
